add LLVM/clang/LLDB C/C++ formatting layer
This change adds a private Spacemacs package called 'llvm-dev'.  When
activated, it pulls in the LLVM C/C++ indentation rules and enables
for any C/C++ file with a buffer name containing 'llvm' or 'lldb'
in the name.

The Emacs indent file is as of LLVM svn trunk r281560.  That file
has a small tweak to enable searching for 'lldb' in addition to
llvm.  This is useful for developing LLDB on macOS, where the Xcode
build dictates that the llvm tree is embedded in the lldb tree.  Thus,
editing lldb files generally doesn't have llvm in the path as a matter
of course.

This change also activates the erc (IRC) layer.