remove semantic, add reveal.js org export, defer org-mode setup
* remove semantic

semantic appears to be slowing down elisp typing, particularly in comments
where, combined with auto completion, it repeatedly explodes archives to do
some elisp code completion (within the comment).  The delays for this errant
behavior are on the order of 10 seconds on my setup.

* add reveal.js exporter for org-mode

Currently I'm expecting the usage file to add the following to the org
doc somewhere:

```
```

* defer org-mode configuration

I have some org-mode code block language setup configuration and other minor
settings that don't need to be set until org is loaded.